Refactor CLI commands into modules
Move infra and init logic into src/commands modules and make main.rs a
thin dispatcher. Keep existing behavior and tests intact while
reorganizing helpers and unit tests.

Closes #70
