+++ src/commands/app.rs
@@ -1,0 +1,7 @@
+pub(crate) fn run_app_add() {
+    println!("bootroot app add: not yet implemented");
+}
+
+pub(crate) fn run_app_info() {
+    println!("bootroot app info: not yet implemented");
+}
