+++ src/commands/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod app;
+pub(crate) mod infra;
+pub(crate) mod init;
+pub(crate) mod status;
+pub(crate) mod verify;
