+++ src/commands/status.rs
@@ -1,0 +1,3 @@
+pub(crate) fn run_status() {
+    println!("bootroot status: not yet implemented");
+}
