+++ src/commands/verify.rs
@@ -1,0 +1,3 @@
+pub(crate) fn run_verify() {
+    println!("bootroot verify: not yet implemented");
+}
